MongoDB Example - Split Models: Client and Token model removed from UserModel to their own files.
