+++ src/examples/mongodb/model/index.ts
@@ -1,1 +1,3 @@
 export * from './UserModel';
+export * from './TokenModel';
+export * from './ClientModel';
